csvJoin Update 0-01
Improved workflow by no longer needing to specify a panel type.
Now supports multiple panels by appending IPID to the output file.

Known Issues : In order for the IPID to be appended, and .sgd file must be linked to the panel within Simpl Windows. Working on a fix for this.

Sometimes the script will catch the panels physical buttons object and parse out the join for that. It might be useful. Still deciding if this is a bug or a feature 👯 .

Currently working on testing/including more panel types.